backend/requirements.txt
fastapi==0.110.0
motor==3.3.2
pymongo==4.6.3
pydantic==2.7.0
python-dotenv==1.0.0
uvicorn[standard]==0.29.0
python-dateutil==2.9.0
